Added training data

+++ utils/parse_pdfs.py
@@ -3,8 +3,8 @@ from PyPDF2 import PdfReader
 import spacy
 import re
 
-pdfs_folder = 'scraped_pdfs/'
-text_folder = 'tokenized_texts/'
+pdfs_folder = 'scraped_train_pdfs/'
+text_folder = 'tokenized_texts/train/'
 
 filenames = os.listdir(pdfs_folder)
 nlp = spacy.load("en_core_web_sm")
@@ -15,13 +15,20 @@ for filename in filenames:
         filename = os.path.join(pdfs_folder, filename)
         textfile = os.path.join(text_folder, filetitle+'.txt')
 
-        reader = PdfReader(filename)
-        print("Number of pages = {}".format(len(reader.pages)))
+        if os.path.exists(textfile):
+            print('skipping since already exists')
+            continue
+
+        try:
+            reader = PdfReader(filename)
+            print("Number of pages = {}".format(len(reader.pages)))
+        except:
+            print("Error parsing PDF")
+            continue
 
         with open(textfile, 'w') as fout:
             for page in reader.pages:
                 text = page.extract_text()
-                # text = text.encode("ascii", "ignore").decode()
                 text = re.sub(r"-+ *\n+ *", "", text)
                 text = re.sub("\n", " ", text)
